app/requirements.txt
streamlit
langchain-google-genai
langchain
langchain-community
langchain-openai
langchain-groq
openai
faiss-cpu
unstructured
pymupdf
langchain-huggingface
transformers
sentencepiece
sentence-transformers
pdfplumber
pandas
tqdm
python-dotenv
streamlit-mic-recorder
SpeechRecognition
gTTS
torch
pydub
streamlit-mic-recorder